'133' fixing versions..

+++ update.sh
@@ -7,13 +7,26 @@ fi
 auth=`cat /root/.gitAuth`
 
 # Functions
-function pushFTP {
-	# Release to FTP
-	version=`cat dsapp-test.sh | grep -wm 1 "dsappversion" | cut -f2 -d"'"`;
-	version=$((version+1))
-	version=`printf "'$version'"`
-	sed -i "s|dsappversion=.*|dsappversion=$version|g" dsapp-test.sh;
-	
+function incrementBuild {
+	clear; echo
+	while true; do
+		read -p "Increment version? [y/n] " ans
+		case $ans in
+			y | Y | yes) makePublic=true; break ;;
+			n | N | no) makePublic=false; break ;;
+			*) echo -e "Invalid entry.\n" ;;
+		esac
+	done
+	if [[ $makePublic ]]; then
+		# Release to FTP
+		version=`cat dsapp-test.sh | grep -wm 1 "dsappversion" | cut -f2 -d"'"`;
+		version=$((version+1))
+		version=`printf "'$version'"`
+		sed -i "s|dsappversion=.*|dsappversion=$version|g" dsapp-test.sh;
+	fi
+}
+
+function uploadFTP {
 	cp dsapp-test.sh dsapp.sh;
 	tar -czf dsapp.tgz dsapp.sh;
 	ftp ftp.novell.com -a <<EOF
@@ -22,12 +35,20 @@ function pushFTP {
 	ha
 	put dsapp.tgz
 EOF
+	if [ $? -ne 0 ]; then
+		echo "Problem uploading to ftp://ftp.novell.com..."
+		return 1
+	fi
 	echo -e "\nCopying to root@tharris7:/wrk/outgoing: "
 	scp dsapp.tgz root@tharris7.lab.novell.com:/wrk/outgoing
+	if [ $? -ne 0 ]; then
+		echo "Problem uploading to tharris7.lab.novell.com:/wrk/outgoing..."
+		return 2
+	fi
 	rm dsapp.sh dsapp.tgz;
-	echo -e "-----------------------------------------"
+	echo -e "\n-----------------------------------------"
 	echo -e "Added to FTP Successfully!"
-	echo "-----------------------------------------"
+	echo -e "-----------------------------------------\n"
 }
 
 function githubPush {
@@ -37,10 +58,6 @@ function githubPush {
 		# Upload to Github.com
 		echo -e "\nUpload to Github.com:"
 		git add dsapp-test.sh update.sh 2> /dev/null
-		gitStatusDsapp=`git status | grep dsapp`;
-		if [[ $gitStatusDsapp && $makePublic ]]; then
-			pushFTP
-		fi
 		if [ $? -eq 0 ]; then
 			#prompt for commit message
 			read -ep "Commit message? " message
@@ -74,7 +91,20 @@ function githubPush {
 		git stash pop
 		echo "Up-to-date. Please upload again."
 	fi
+}
 
+function newPublicRelease {
+	incrementBuild
+	uploadFTP
+	read -p "[Exit]";
+	exit 0
+}
+
+function newInternalRelease {
+	incrementBuild
+	githubPush
+	read -p "[Exit]";
+	exit 0
 }
 
 #Menu loop
@@ -84,27 +114,19 @@ do
 echo '     
 	SCRIPT MENU      
    '
- echo -e "\t1. Make Public (FTP + Github)"
- echo -e "\t2. Push to Github"
+ echo -e "\t1. Public Release (FTP)"
+ echo -e "\n\t2. Push to Github"
  echo -e "\t3. Pull from Github"
  echo -e "\n\t0. Quit"
  echo -n -e "\n\tSelection: "
  read opt;
  case $opt in
 
- 1)	makePublic=true;
-	githubPush
-	echo -e "Successful Upload!";
-	read -p "[Exit]";
-	exit 0
+ 1)	newPublicRelease
 	;;
 
  2) # Push to Github Only
-	makePublic=false;
-	githubPush
-	echo -e "Successful Upload!";
-	read -p "[Exit]";
-	exit 0
+	newInternalRelease
 	;;
 
  3)	# Pull from Github
